% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_file_names}
\alias{get_file_names}
\title{get_file_names}
\usage{
get_file_names(studyid, file_dir = "data-raw/trees/treebase")
}
\arguments{
\item{studyid}{TreeBASE study id}

\item{file_dir}{Character vector of a directory with one or more tree files}
}
\description{
Write nexus or newick files to pdf or png.
}
