% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_tip_values}
\alias{get_tip_values}
\title{Assign tip values to a tree based on OTU information file.}
\usage{
get_tip_values(treefile, otufile)
}
\arguments{
\item{treefile}{A character vector with the path to the tree you want to plot in newick format.}

\item{otufile}{A character vector with the path to the OTU information csv file generated by Physcraper}
}
\value{
a list of a phylo object and a vector of tip values
}
\description{
Assign tip values to a tree based on OTU information file.
}
\examples{
treefile = 'data/pg_2827_tree6577/run_pg_2827tree6577_run4/RAxML_bestTree.2020-07-31'
otufile = 'data/pg_2827_tree6577/outputs_pg_2827tree6577/otu_info_pg_2827tree6577.csv'
}
\author{
Emily Jane McTavish
}
