% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_branches}
\alias{plot_branches}
\title{Get a plot of a tree with tips and branches colored following a tip value.}
\usage{
plot_branches(treefile, otufile, ...)
}
\arguments{
\item{treefile}{A character vector with the path to the tree you want to plot in newick format.}

\item{otufile}{A character vector with the path to the OTU information csv file generated by Physcraper}
}
\value{
a plot
}
\description{
Get a plot of a tree with tips and branches colored following a tip value.
}
