% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_branches_method2}
\alias{plot_branches_method2}
\title{Plot a tree with ggtree branches and tips colored according to molecular data, method 2
It is giving an error with ggtree, make an issue or contact the author.}
\usage{
plot_branches_method2(
  x,
  tip_label = "otu",
  drop_outgroup = TRUE,
  ladderize_tree = TRUE,
  color = "red",
  ...
)
}
\arguments{
\item{x}{A list from get_tip_values}

\item{tip_label}{A character vector. Can be one of "otu" or "taxon"}

\item{drop_outgroup}{Boolean}

\item{ladderize_tree}{Boolean}
}
\value{
a plot
}
\description{
Plot a tree with ggtree branches and tips colored according to molecular data, method 2
It is giving an error with ggtree, make an issue or contact the author.
}
\author{
Luna L. Sanchez Reyes
}
