% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sum_tips}
\alias{sum_tips}
\title{Cummulative sum of tip values for each node/branch down to the root}
\usage{
sum_tips(tree, values)
}
\arguments{
\item{tree}{a phylo object}

\item{values}{named numeric character vector with names corresponding to tip labels}
}
\value{
a vector
}
\description{
Cummulative sum of tip values for each node/branch down to the root
}
\author{
Luna L. Sanchez Reyes
}
