% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{trenex2pdf}
\alias{trenex2pdf}
\title{trenex2pdf}
\usage{
trenex2pdf(
  treefile,
  treefilesdir = "../data-raw/trees/treebase/S1091/",
  treepdfsdir = "../data-raw/trees/treebase/S1091-pdf/",
  height = 7,
  ...
)
}
\arguments{
\item{treefile}{A character vector with the name of the nexus or newick file}

\item{studyid}{TreeBASE study id}
}
\description{
trenex2pdf
}
\details{
Write nexus tree files to pdf.
}
